requirements-dev.txt
django-debug-toolbar
ruff